Implemented AbstractCommandContainer
