+++ src/Command/Play/CommandContainer.php
@@ -2,10 +2,10 @@
 
 namespace ChessServer\Command\Play;
 
-class CommandContainer
-{
-    private $obj;
+use ChessServer\Command\AbstractCommandContainer;
 
+class CommandContainer extends AbstractCommandContainer
+{
     public function __construct()
     {
         $this->obj = new \SplObjectStorage;
@@ -23,32 +23,5 @@ class CommandContainer
         $this->obj->attach(new StockfishCommand());
         $this->obj->attach(new TakebackCommand());
         $this->obj->attach(new UndoCommand());
-    }
-
-    public function findByName(string $name)
-    {
-        $this->obj->rewind();
-        while ($this->obj->valid()) {
-            if ($this->obj->current()->name === $name) {
-                return $this->obj->current();
-            }
-            $this->obj->next();
-        }
-
-        return null;
-    }
-
-    public function help()
-    {
-        $o = '';
-        $this->obj->rewind();
-        while ($this->obj->valid()) {
-            $o .= $this->obj->current()->name;
-            $this->obj->current()->params ? $o .= ' ' . json_encode($this->obj->current()->params) : null;
-            $o .= ' ' . $this->obj->current()->description . PHP_EOL;
-            $this->obj->next();
-        }
-
-        return $o;
     }
 }
